requirements.txt
aiohttp==3.7.3
attrs==20.3.0
colorlog==4.6.2
multidict==5.1.0
C:/Users/Lukec/PycharmProjects/pythonProject/marshie
